# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Restaurant",
        "description": "REST API para autenticación de usuarios, administración del menú y gestión de ingredientes.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Lista las órdenes del usuario autenticado o todas si es admin",
                "operationId": "745ae49c58e1f04df4805e29759b70ce",
                "parameters": [
                    {
                        "name": "all",
                        "in": "query",
                        "description": "Solo admins: incluir todas las órdenes",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de órdenes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderCollectionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Crea una nueva orden",
                "operationId": "128abc05f8e101791ba28a1c7136f23c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "address_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "address": {
                                        "properties": {
                                            "address_line": {
                                                "type": "string",
                                                "maxLength": 200
                                            },
                                            "city_id": {
                                                "type": "integer"
                                            },
                                            "sector": {
                                                "type": "string",
                                                "maxLength": 100,
                                                "nullable": true
                                            },
                                            "reference": {
                                                "type": "string",
                                                "maxLength": 200,
                                                "nullable": true
                                            }
                                        },
                                        "type": "object",
                                        "nullable": true
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "pizza_id"
                                            ],
                                            "properties": {
                                                "pizza_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "default": 1,
                                                    "minimum": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtiene una orden específica",
                "operationId": "6194b0f9dd0dd8207f4ed56a92042dc0",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orden encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{order}/status": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Actualiza el estado de una orden",
                "operationId": "82ff70d2bfe3fd5d8d9f8d9ab3703460",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "paid",
                                            "preparing",
                                            "delivered",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{order}/payments": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Registra un pago para la orden",
                "operationId": "9dbce8e29a06a9d7f2ebe358e563093f",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0.01
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "completed",
                                            "failed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pago registrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PizzaImage": {
                "properties": {
                    "public_id": {
                        "type": "string"
                    },
                    "secure_url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "format": {
                        "type": "string",
                        "example": "jpg"
                    },
                    "resource_type": {
                        "type": "string",
                        "example": "image"
                    },
                    "bytes": {
                        "type": "integer"
                    },
                    "width": {
                        "type": "integer"
                    },
                    "height": {
                        "type": "integer"
                    }
                },
                "type": "object",
                "nullable": true
            },
            "PizzaResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 99
                    },
                    "name": {
                        "type": "string",
                        "example": "Diavola"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "string",
                        "example": "12.90"
                    },
                    "is_recommended": {
                        "type": "boolean",
                        "example": true
                    },
                    "image_url": {
                        "$ref": "#/components/schemas/PizzaImage"
                    }
                },
                "type": "object"
            },
...
            "AddressResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 9001
                    },
                    "address_line": {
                        "type": "string",
                        "example": "Av. Central #123"
                    },
                    "sector": {
                        "type": "string",
                        "example": "Zona Colonial",
                        "nullable": true
                    },
                    "reference": {
                        "type": "string",
                        "example": "Frente al parque",
                        "nullable": true
                    },
                    "city_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "OrderItemResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1001
                    },
                    "pizza_id": {
                        "type": "integer",
                        "example": 15
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "unit_price": {
                        "type": "string",
                        "example": "12.90"
                    },
                    "pizza": {
                        "$ref": "#/components/schemas/PizzaResource"
                    }
                },
                "type": "object"
            },
            "PaymentResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 77
                    },
                    "amount": {
                        "type": "string",
                        "example": "25.80"
                    },
                    "status": {
                        "type": "string",
                        "example": "completed"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "OrderResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5000
                    },
                    "total": {
                        "type": "string",
                        "example": "25.80"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "address": {
                        "$ref": "#/components/schemas/AddressResource"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItemResource"
                        }
                    },
                    "payments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PaymentResource"
                        }
                    }
                },
                "type": "object"
            },
            "OrderResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "order": {
                                "$ref": "#/components/schemas/OrderResource"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "OrderCollectionResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "orders": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PaymentResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "payment": {
                                "$ref": "#/components/schemas/PaymentResource"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "Las credenciales no coinciden."
                            ]
                        },
                        "additionalProperties": {
                            "schema": "Schemas",
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "UnauthorizedError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    },
                    "status": {
                        "type": "integer",
                        "example": 401
                    }
                },
                "type": "object"
            },
            "ForbiddenError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    },
                    "status": {
                        "type": "integer",
                        "example": 403
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Incluye el token generado en el login o registro usando el formato: Bearer {token}",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}